🚚 Rename: s3 관련 기능의 폴더를 생성
